requirements.txt
openai==2.16.0
pillow==12.1.0
python-dotenv==1.2.1
streamlit==1.53.1
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
transformers==5.0.0